fix cron update positions

+++ app/Console/Kernel.php
@@ -96,10 +96,10 @@ class Kernel extends ConsoleKernel
 
             foreach ($engines as $engine) {
 
-                /*if (empty($engine->project)) {
+                if (empty($engine->project)) {
                     $engine->delete();
                     continue;
-                }*/
+                }
 
                 $time = explode(':', $engine->time ?? '00:00');
                 $hour = (int)$time[0];
